+++ lib/fluttered.dart
@@ -1,3 +1,5 @@
 library fluttered;
 
 export 'src/widget/widget.dart';
+export 'src/util/util.dart';
+export 'src/config/config.dart';