+++ lib/src/config/config.dart
@@ -1,0 +1,2 @@
+export 'fluttered_config.dart';
+export 'theme_config.dart';