+++ lib/src/config/fluttered_config.dart
@@ -1,0 +1,8 @@
+import 'package:fluttered/src/config/theme_config.dart';
+
+/// Set Global configurations for fluttered.
+void flutteredConfig({
+  ThemeConfig theme,
+}) {
+  themeConfig = theme;
+}