+++ CHANGELOG.md
@@ -18,4 +18,14 @@ and `StatefulImmutabalePVC` are removed. Use `MutablePVC` and `ImmutablePVC` ins
 **New Features:**
 - StatefulWrapper widget added.
 
+## [0.2.0]
+
+**Breaking Change:**
+- `onInit` parameter changed to `initState` in `StatefulWrapper`.
+
+**New Features:**
+-`ThemeManager` widget added.
+- `didUpdateWidget` and `didChangeDependencies` parameter added to `StatefulWrapper`.
+- Utility methods to calculate `width` and `height` based on `screenSize` added.
+